Move camera controler to its own subpackage
