+++ control/adafruithat/planktoscope/camera/__init__.py
@@ -1,0 +1,1 @@
+"""camera provides low-level functionality for camera control."""
