+++ control/adafruithat/planktoscope/camera/mjpeg.py
