+++ control/pscopehat/planktoscope/camera/mjpeg.py
